+++ src/Components/Blog/Blog.js
@@ -1,10 +1,14 @@
 import React from 'react'
-
+import test from '../../assets/test.jpg'
 function Blog() {
     return (
         <div>
             <h1>Blog</h1>
+            
+           <p>Lorem ipsum dolor sit amet. Lorem ipsum dolor, sit amet consectetur adipisicing elit. Commodi quasi quibusdam suscipit deserunt, quia atque rerum. Assumenda nam porro quo ullam similique ratione, officiis soluta fugiat. Facilis at, perferendis nobis ad, quidem veniam esse culpa, nostrum voluptates libero aperiam eligendi. Fuga quibusdam nam officiis, nihil quam deleniti odit earum necessitatibus.</p>
+           <img style={{width:"500px"}} src={test} alt="" />
         </div>
+
     )
 }