+++ src/Components/Content/Sub-Content/Video.js
@@ -1,9 +1,10 @@
 import React from 'react'
-
+import test2 from '../../../assets/test2.jpg'
 function Video() {
     return (
         <div>
             <h3>Video under content</h3>
+            <img style={{width:"500px"}} src={test2} alt="" />
         </div>
     )
 }