+++ src/Components/MenuItem.js
@@ -1,20 +1,21 @@
 import React, { useState } from 'react';
+import { NavLink } from 'react-router-dom';
 
 function MenuItem(props) {
-    const { name, subMenus, iconClassName, to, onClick } = props;
+    const { name, subMenus, iconClassName, to, onClick, exact } = props;
     const [expandSubMenu, setExpandSubMenu] = useState(false)
     return (
         <li onClick={onClick}>
-            <button onClick={() => setExpandSubMenu(!expandSubMenu)} className="menu-item">
+            <NavLink exact={exact} activeClassName="selected-main-menu" to={to} onClick={() => setExpandSubMenu(!expandSubMenu)} className="menu-item">
                 <div className="menu-icon"> <i className={iconClassName}></i> </div>
                 <span> {name}</span>
-            </button>
+            </NavLink >
             {
                 subMenus && subMenus.length > 0 ? <ul className={`sub-menu ${expandSubMenu ? "sub-menu-active" : ""}`}>
                     {
                         subMenus.map((menu, index) =>
                             <li key={index} >
-                                <button className="sub-menu-items">{menu.name}</button>
+                                <NavLink exact={exact} activeClassName="selected-sub-menu" to={menu.to} className="sub-menu-items">{menu.name}</NavLink >
                             </li>)
                     }
                 </ul> : null